fix(fe): useSearchParams를 클라이언트 컴포넌트로 분리
- useSearchParams로 인한 빌드 에러 해결을 위해 CSR 전용 컴포넌트로 분리
